feat: ajustado menu do sistema para modelo atômico
